+++ src/js/contato.js
@@ -1,3 +1,11 @@
+// Função para carregar o webhookURL do arquivo .json
+async function getWebhookURL() {
+    const response = await fetch('webhook.json');
+    const data = await response.json();
+    return data.webhookURL;  // Retorna a URL do webhook
+}
+
+// Função para enviar os dados ao Discord
 document.getElementById("contact-form").addEventListener("submit", function(event) {
     event.preventDefault();
 
@@ -10,31 +18,65 @@ document.getElementById("contact-form").addEventListener("submit", function(even
     const regex = /^\(\d{2}\)\s?\d{5}-\d{4}$/;
     if (!regex.test(telefone)) {
         alert("Por favor, insira um número de telefone válido com o formato: (XX) XXXXX-XXXX");
-        return; // Impede o envio do formulário se o telefone for inválido
+        return;
     }
 
-    // Dados do formulário
-    const formData = {
-        nome,
-        email,
-        telefone,
-        mensagem
+    // Criando o Embed com os dados
+    const embed = {
+        "embeds": [
+            {
+                "title": "<a:obsidian:1229497555000234138> Contato Recebido",
+                "description": `<:obsidian:1229727889268674641> **Nome**: ${nome}\n`,
+                "color": 0x9e26d6,
+                "footer": {
+                    "text": "Atenciosamente Obsidian Codes™",
+                    "icon_url": "https://cdn.discordapp.com/attachments/1281996602743066664/1343818881952776255/home.png"
+                },
+                "thumbnail": {
+                    "url": "https://cdn.discordapp.com/attachments/1281996602743066664/1343818881952776255/home.png"
+                },
+                "fields": [
+                    {
+                        "name": "<:obsidian:1229727889268674641> Detalhes do contato",
+                        "value": `Email: ${email}\nTelefone: ${telefone}`,
+                        "inline": true
+                    },
+                    {
+                        "name": "Mensagem",
+                        "value": mensagem,
+                        "inline": false
+                    }
+                ],
+                "actions": [
+                    {
+                        "type": 1,
+                        "label": "Responder",
+                        "url": "https://discord.com/channels/1229106476799234180/1229106514829246494"
+                    }
+                ]
+            }
+        ]
     };
 
-    // Enviar para a função serverless no Netlify
-    fetch('/.netlify/functions/sendWebhook', {
-        method: 'POST',
-        headers: {
-            'Content-Type': 'application/json'
-        },
-        body: JSON.stringify(formData)
-    })
-    .then(response => response.json())
-    .then(data => {
-        alert(data.message); // Mostra a mensagem de sucesso ou erro
-        document.getElementById("contact-form").reset();  // Limpa o formulário
-    })
-    .catch(error => {
-        alert('Erro ao enviar a mensagem.');
+    // Obter o webhookURL de forma assíncrona e enviar os dados
+    getWebhookURL().then(webhookURL => {
+        fetch(webhookURL, {
+            method: "POST",
+            headers: {
+                "Content-Type": "application/json"
+            },
+            body: JSON.stringify(embed)
+        })
+        .then(response => {
+            if (response.ok) {
+                alert("Mensagem enviada com sucesso para o Discord!");
+                document.getElementById("contact-form").reset(); // Limpa o formulário
+            } else {
+                alert("Erro ao enviar a mensagem.");
+            }
+        })
+        .catch(() => {
+            alert("Erro ao enviar a mensagem.");
+        });
     });
 });
